requirements.txt
tqdm
openai>=1.40.0
anthropic>=0.39.0
google-generativeai>=0.8.3
python-dotenv>=1.0.1
